+++ src/TRegx/CleanRegex/Match/Details/Group/ReplaceNotMatchedGroup.php
@@ -1,7 +1,7 @@
 <?php
 namespace TRegx\CleanRegex\Match\Details\Group;
 
-class ReplaceNotMatchedGroup extends NotMatchedGroup implements ReplaceDetailGroup
+class ReplaceNotMatchedGroup extends NotMatchedGroup implements ReplaceDetailGroup, ReplaceMatchGroup
 {
     public function modifiedOffset(): int
     {